// catalog_service/src/utils/broker/broker.type.ts
import { MessageType, CatalogEvents, TOPIC_TYPE } from "../../types";

export interface PublishType {
  headers: Record<string, any>;
  topic: TOPIC_TYPE;
  event: CatalogEvents;
  message: Record<string, any>;
}

export type MessageHandler = (input: MessageType) => void | Promise<void>;

export type MessageBrokerType = {
  connectProducer: <T>() => Promise<T>;
  disconnectProducer: () => Promise<void>;
  publish: (data: PublishType) => Promise<boolean>;

  connectConsumer: <T>() => Promise<T>;
  disconnectConsumer: () => Promise<void>;
  subscribe: (messageHandler: MessageHandler, topic: TOPIC_TYPE) => Promise<void>;
};